refactor: moved text input and button to uiComponents
